종자원관리 예외사항 수정하였습니다.

+++ src/main/java/com/digitalresource/Entity/Feature.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Feature {
     private int feature_id;
-    private int feature_num;
+    private Double feature_num;
     private String feature_name;
     private String feature_expression;
     private String feature_level;
